aiagents4pharma/talk2scholars/requirements.txt
ipykernel==6.29.5
ipython==8.31.0
langchain-openai==0.3.0
langgraph==0.2.62
openai==1.59.6
pandas==2.2.3
pydantic==2.10.5
requests==2.32.3
streamlit==1.41.1
tabulate==0.9.0
requests
pandas
openai
langchain_openai
langgraph
typing_extensions
capture
langchain_community 
langchain_anthropic 
langchain_experimental
hydra-core